Fix import of ovh_ip_reverse resources with IPv6
When importing an `ovh_ip_reverse` resources, the user is expected to
provide the IP block and IP address separated with a colon, e.g.
`198.51.100.0/24:198.51.100.42`:

  * `198.51.100.0/24` is extracted as the IP block;
  * `198.51.100.42` is extracted as the address.

Unfortunately, the `:` separator does not work well with IPv6 addresses,
e.g. `2001:0DB8::/32:2001:0DB8::42` is interpreted as:

  * `2001` is extracted as the IP block;
  * `0DB8::/32:2001:0DB8::42` is extracted as the address.

This is obviously wrong and raise an API error:

```
│ Error: calling /ip/%222001/reverse/0DB8::%2F32:2001:0DB8::42:
│ 	 HTTP Error 400: "[ipReverse] Given data (0DB8::/32:2001:0DB8::42) is not valid for type ip"
```

Replace the `:` separator with a `|` separator to avoid this issue.